refactor(canvas): deepen chat runtime internals
