+++ apps/canvas-workspace/src/renderer/src/agent-chat/runtime/conversationKey.ts
@@ -1,0 +1,14 @@
+import type { AgentScope } from '../../types';
+import type { ConversationKey } from '../../../../shared/conversation-runtime';
+
+export const conversationKeyFromScope = (
+  scope: AgentScope,
+  sessionId: string,
+): ConversationKey => ({
+  storeId: scope.kind === 'workspace'
+    ? scope.workspaceId
+    : scope.kind === 'scheduled'
+      ? `__scheduled__-${scope.taskId}`
+      : '__global_chat__',
+  sessionId,
+});
