+++ apps/canvas-workspace/src/renderer/src/agent-chat/runtime/useConversationRuntimeStream.ts
@@ -492,16 +492,3 @@ export function useConversationRuntimeStream({
     toggleToolExpand,
   };
 }
-
-/** Build a conversation key from a scope + session id (thin helper). */
-export const conversationKeyFromScope = (
-  scope: AgentScope,
-  sessionId: string,
-): ConversationKey => ({
-  storeId: scope.kind === 'workspace'
-    ? scope.workspaceId
-    : scope.kind === 'scheduled'
-      ? `__scheduled__-${scope.taskId}`
-      : '__global_chat__',
-  sessionId,
-});
