+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessages/index.tsx
@@ -21,7 +21,6 @@ import { localPathFromHref } from '../utils/localFileLinks';
 import { ChatClarificationCard } from '../ChatClarificationCard';
 import { useChatMessagesStatus } from './useChatMessagesStatus';
 import { tabRefFromMentionElement } from '../utils/tabMentions';
-
 /** How close (px) to the bottom still counts as "reading the tail" — within
  *  this band the view keeps following the stream; beyond it the user has
  *  scrolled up to read and auto-scroll must not yank them back. */
